+++ docs/src/content/docs/changelog.md
@@ -18,6 +18,8 @@ Requires libvips v8.17.0
 
 * Expose `keepDuplicateFrames` GIF output parameter.
 
+* Add support for RAW digital camera image input. Requires libvips compiled with libraw support.
+
 * Add `pageHeight` option to `create` and `raw` input for animated images.
   [#3236](https://github.com/lovell/sharp/issues/3236)